requirements.txt
deta==1.1.0
plotly==5.14.1
python-dotenv==1.0.0
streamlit==1.20.0
streamlit_option_menu==0.3.2
